refactor: rename OnwardItem to Trail
